rename branch 'core' -> 'GUI', and Directory Rectification, fix #21
