+++ GUI/android-sources/AndroidManifest.xml
