+++ GUI/android-sources/res/drawable-hdpi/icon.png
