+++ GUI/android-sources/res/drawable-ldpi/icon.png
