+++ GUI/android-sources/res/drawable-mdpi/icon.png
