+++ GUI/android-sources/res/drawable/1/icon_hdpi72x72.png
