+++ GUI/android-sources/res/drawable/1/icon_ldpi36x36.png
