+++ GUI/android-sources/res/drawable/1/icon_mdpi48x48.png
