+++ GUI/android-sources/res/drawable/2/icon_hdpi72x72.png
