+++ GUI/android-sources/res/drawable/2/icon_ldpi36x36.png
