+++ GUI/android-sources/res/drawable/2/icon_mdpi48x48.png
