+++ GUI/android-sources/res/drawable/3/icon_hdpi72x72.png
