+++ GUI/android-sources/res/drawable/3/icon_ldpi36x36.png
