+++ GUI/android-sources/res/drawable/3/icon_mdpi48x48.png
