+++ GUI/android-sources/res/drawable/4/icon_hdpi72x72.png
