+++ GUI/android-sources/res/drawable/4/icon_ldpi36x36.png
