+++ GUI/android-sources/res/drawable/4/icon_mdpi48x48.png
