+++ GUI/android-sources/res/drawable/4/xu.png
