+++ GUI/qml.qrc
