+++ GUI/webchannelobject/webchannelobject.h
