+++ GUI/websocketchannel/websocketclientwrapper.cpp
