+++ GUI/websocketchannel/websocketclientwrapper.h
