+++ GUI/websocketchannel/websockettransport.h
